+++ src/videostreamindex/DefaultVideoStreamIndex.ts
@@ -116,8 +116,11 @@ export default class DefaultVideoStreamIndex implements VideoStreamIndex {
     return set;
   }
 
-  allVideoSendingAttendeesExcludingSelf(selfAttendeeId: string): Set<string> {
-    const attendees = new Set<string>();
+  allVideoSendingAttendeesExcludingSelf(
+    selfAttendeeId: string
+  ): { attendeeId: string; externalUserId: string }[] {
+    const attendees = [];
+    const attendeeSet = new Set<string>();
     if (this.currentIndex) {
       if (this.currentIndex.sources.length) {
         for (const stream of this.currentIndex.sources) {
@@ -125,12 +128,15 @@ export default class DefaultVideoStreamIndex implements VideoStreamIndex {
             stream.attendeeId !== selfAttendeeId &&
             stream.mediaType === SdkStreamMediaType.VIDEO
           ) {
-            attendees.add(stream.attendeeId);
+            const { attendeeId, externalUserId } = stream;
+            if (!attendeeSet.has(attendeeId)) {
+              attendees.push({ attendeeId, externalUserId });
+              attendeeSet.add(attendeeId);
+            }
           }
         }
       }
     }
-
     return attendees;
   }