+++ src/videotile/VideoTileStatusCode.ts
@@ -1,0 +1,9 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+enum VideoTileStatusCode {
+  VideoEnded,
+  VideoRemovedDueToPolicy,
+}
+
+export default VideoTileStatusCode;